Separate out the tool classes to avoid polluting the final JAR
